requirements.txt
boto3 == 1.9.25
PyYAML == 3.13
M2Crypto == 0.30.1